Refactored the ServerWrapper class.
Instead of doing the if (server == null) checks, the ServerWrapper now
uses ServerStates. The two states are InactiveServerState and
ActiveServerState.

I also added some logging here and there.
Also, if a server fails to get created, it is re-attempted several times
before actually throwing a fatal exception.

The TestServerWrapper is now completely out of date, so every test there
is removed.